+++ jest.config.ts
@@ -7,6 +7,14 @@ import type { Config } from 'jest';
  */
 const config: Config = {
     preset: 'ts-jest',
+    transform: {
+        '^.+\\.tsx?$': ['ts-jest', {
+            tsconfig: {
+                module: 'CommonJS',
+                moduleResolution: 'Node',
+            },
+        }],
+    },
     testEnvironment: 'node',
     clearMocks: true,
     coverageProvider: 'v8',
@@ -17,9 +25,17 @@ const config: Config = {
     collectCoverageFrom: [
         'src/**/*.ts',
         '!src/types.ts',
-        '!src/index.ts',
         '!src/examples/**'
     ],
+
+    coverageThreshold: {
+        global: {
+            statements: 95,
+            lines: 95,
+            branches: 90,
+            functions: 90,
+        },
+    },
     
     modulePathIgnorePatterns: [
         '<rootDir>/dist/'